# api/api.yaml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: TorrPlay API
  description: A torrent streaming application with memory-managed piece storage.
  termsOfService: https://example.com/terms/
  contact:
    name: TorrPlay Team
    email: support@example.com
    url: https://github.com/torrplay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
security:
  - basicAuth: []
  - bearerAuth: []
...
paths:
...
  /api/v1/torrents/backup:
    get:
      tags:
        - v1
      summary: Backup torrents and posters.
      description: Backup all torrents and posters to a single file.
      operationId: backupTorrents
      responses:
        "200":
          description: Backup file.
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/Backup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/torrents/restore:
    post:
      tags:
        - v1
      summary: Restore torrents and posters.
      description: Restore torrents and posters from a backup file.
      operationId: restoreTorrents
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "204":
          description: Restored successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/torrents/{hash}:
    get:
      tags:
        - v1
      summary: Get torrent metadata.
      description: Retrieve detailed information about a specific torrent by its info hash.
      operationId: getTorrent
      parameters:
        - $ref: "#/components/parameters/Hash"
      responses:
        "200":
          description: Torrent retrieved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Torrent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"

    patch:
      tags:
        - v1
      summary: Update torrent metadata.
      description: Update metadata for an existing torrent. Only modifiable fields can be updated.
      operationId: updateTorrent
      parameters:
        - $ref: "#/components/parameters/Hash"
      requestBody:
        description: Torrent metadata to update.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TorrentUpdate"
      responses:
        "200":
          description: Torrent updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Torrent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - v1
      summary: Delete a torrent.
      description: Remove a torrent from the system. This will also free any memory used by the torrent's pieces.
      operationId: deleteTorrent
      parameters:
        - $ref: "#/components/parameters/Hash"
      responses:
        "204":
          description: Torrent deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    Hash:
      name: hash
      in: path
      description: Torrent info hash.
      required: true
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{40}$"
        example: "08ada5a7a6183aae1e09d831df6748d566095a10"
        x-go-type: metainfo.Hash
        x-go-type-import:
          path: github.com/anacrolix/torrent/metainfo
...
  responses:
...
    BadRequest:
      description: The request was invalid or cannot be served.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Invalid info hash format."
            code: 400
            details: "Info hash must be 40 hex characters."
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized."
            code: 401
            details: "Authentication token is missing or invalid."
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Torrent not found."
            code: 404
            details: "No torrent found with hash: da39a3ee5e6b4b0d3255bfef95601890afd80709."
...
    InternalServerError:
      description: An internal server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error."
            code: 500
            details: "Failed to allocate memory for piece storage."
...
    GatewayTimeout:
      description: The server did not receive a timely response from an upstream server.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Gateway timeout."
            code: 504
            details: "Timeout waiting for torrent metadata."
  schemas:
...
    Backup:
      type: object
      required:
        - torrents
        - posters
      properties:
        torrents:
          type: array
          items:
            $ref: "#/components/schemas/Torrent"
        posters:
          type: object
          additionalProperties:
            type: string
            format: binary
    Error:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
          description: Human-readable error message.
          example: "Torrent not found"
        code:
          type: integer
          description: HTTP status code.
          minimum: 400
          maximum: 599
          example: 404
...
    TorrentStorage:
      type: string
      description: Use memory or file storage.
      enum:
        - "memory"
        - "file"
      example: "memory"
...
    Torrent:
      allOf:
        - $ref: "#/components/schemas/TorrentAdd"
        - $ref: "#/components/schemas/TorrentUpdate"
        - type: object
          required:
            - hash
            - magnet
            - name
            - files
            - total_size
            - piece_count
          properties:
            files:
              type: array
              description: List of files contained in the torrent.
              items:
                $ref: "#/components/schemas/TorrentFile"
            hash:
              type: string
              x-go-type: metainfo.Hash
              x-go-type-import:
                path: github.com/anacrolix/torrent/metainfo
              description: SHA-1 info hash of the torrent (40 hex characters).
              pattern: "^[0-9a-fA-F]{40}$"
              example: "08ada5a7a6183aae1e09d831df6748d566095a10"
            name:
              type: string
              description: Original name of the torrent from metadata.
              example: "Sintel"
            total_size:
              type: integer
              format: int64
              description: Total size of all files in bytes.
              example: 3711950848
            piece_count:
              type: integer
              description: Total number of pieces in the torrent.
              example: 1418
            piece_size:
              type: integer
              format: int64
              description: Standard piece size in bytes.
              example: 262144
            created_at:
              type: string
              format: date-time
              description: When the torrent was added to the system.
              example: "2024-01-15T10:30:00Z"
            updated_at:
              type: string
              format: date-time
              description: Last time the torrent was modified.
              example: "2024-01-15T11:45:00Z"
    TorrentAdd:
      type: object
      description: Input for adding a new torrent.
      properties:
        category:
          type: string
          description: Category for organizing torrents.
          example: "Movies"
        hash:
          type: string
          x-go-type: metainfo.Hash
          x-go-type-import:
            path: github.com/anacrolix/torrent/metainfo
          description: SHA-1 info hash of the torrent (40 hex characters).
          pattern: "^[0-9a-fA-F]{40}$"
          example: "08ada5a7a6183aae1e09d831df6748d566095a10"
        magnet:
          type: string
          description: Magnet URI.
          pattern: '^magnet:\?xt=urn:btih:([a-fA-F0-9]{40}|[a-zA-Z2-7]{32})'
          example: "magnet:?xt=urn:btih:08ada5a7a6183aae1e09d831df6748d566095a10&dn=Sintel"
        poster:
          type: string
          description: URL or data URI to poster image.
          example: "https://example.com/poster.jpg"
        storage:
          allOf:
            - $ref: "#/components/schemas/TorrentStorage"
          description: Use memory or file storage.
          default: "memory"
        title:
          type: string
          description: Display title for the torrent.
          minLength: 1
          maxLength: 1024
          example: "Sintel"
...
    TorrentFileUpdate:
      type: object
      description: "Metadata for updating a file within a torrent."
      required:
        - path
        - viewed
      properties:
        path:
          type: string
          description: The full path of the file within the torrent.
        viewed:
          type: boolean
          description: Set to true to mark as viewed, false to mark as unviewed.
    TorrentUpdate:
      type: object
      description: Updateable torrent metadata.
      properties:
        category:
          type: string
          description: Category for organizing torrents.
          example: "Movies"
        files:
          type: array
          description: List of files to update within the torrent.
          items:
            $ref: "#/components/schemas/TorrentFileUpdate"
          readOnly: true
        poster:
          type: string
          description: URL or data URI to poster image.
          example: "https://example.com/new-poster.jpg"
        storage:
          $ref: "#/components/schemas/TorrentStorage"
        title:
          type: string
          description: Display title for the torrent.
          minLength: 1
          maxLength: 1024
          example: "Sintel"
    TorrentFile:
      type: object
      description: File within a torrent.
      required:
        - length
        - name
        - path
      properties:
        length:
          type: integer
          format: int64
          description: File size in bytes.
          minimum: 0
          example: 129241752
        name:
          type: string
          description: Base name of the file.
          example: "Sintel.mp4"
        path:
          type: string
          description: Full path of the file within the torrent.
          example: "Sintel/Sintel.mp4"
        viewed_at:
          type: string
          format: date-time
          description: When the file was last viewed.
          example: "2024-01-15T11:45:00Z"
          nullable: true
...